+++ src/examples/index.ts
@@ -1,2 +1,2 @@
 import command from '../';
-command(`${ __dirname }/commands/**`);
+command(`${__dirname}/**/*.cmd.js`);