+++ typings/index.d.ts
@@ -1,6 +1,0 @@
-/// <reference path="globals/mocha/index.d.ts" />
-/// <reference path="globals/node/index.d.ts" />
-/// <reference path="modules/chai/index.d.ts" />
-/// <reference path="modules/chalk/index.d.ts" />
-/// <reference path="modules/minimist/index.d.ts" />
-/// <reference path="modules/ramda/index.d.ts" />